requirements.txt
tweepy
pandas
s3fs
apache-airflow
PyYAML